// src/lib/weebdex/model/chapter.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * WeebDex API
 * WeebDex is an ad-free, scanlator-first online manga reader providing high-quality, zero-compression images. <br /><br />
Contact: [delta@weebdex.org](mailto:delta@weebdex.org) or join our [Discord Server](https://discord.gg/syp3a6Eucq).<br /><br />
All API endpoints have a global rate limit of 5 requests per second per IP. <br />
To avoid future issues, include the Origin: https://weebdex.org and Referer: https://weebdex.org/ headers when making API requests.<br/>
 * OpenAPI spec version: 1.2.0
 */
import type { ChapterRelationships } from "./chapterRelationships";
import type { Page } from "./page";

export interface Chapter {
  chapter?: string;
  created_at?: string;
  data?: Page[];
  data_optimized?: Page[];
  deleted_at?: string;
  id?: string;
  is_unavailable?: boolean;
  language?: string;
  node?: string;
  published_at?: string;
  relationships?: ChapterRelationships;
  source_id?: string;
  title?: string;
  updated_at?: string;
  version?: number;
  volume?: string;
}